finish the partition repair across the remaining packages
The crate partition left compile and formatting drift outside the
fourteen domain packages: the neural-net, visual, parsing, decode, text,
reduce, builder, bench, and aot packages each carried imports, module
declarations, or test material that named a pre-partition path.

The tree is now uniformly formatted. Formatting was the first failing step
of every platform job, which skipped every check behind it, so no other
result in this campaign was observable until it passed.

+++ vyre-libs-nn/src/nn/linear/layer/linear_4bit/grouped_layout.rs
@@ -44,7 +44,7 @@ pub(super) fn affine_grouped_buffers(
     sidecar_count: u32,
     out_dim: u32,
     padded_output_count: u32,
-    output_byte_len: usize,
+    output_byte_len: u64,
 ) -> Vec<BufferDecl> {
     let [x, w_packed, scale, zero_point, b, out] = names;
     vec![
